requirements.txt
lxml==4.5.1
Pillow==6.2.2
python-docx==0.8.10